refactor(#21): BaseEntity 패키지 구조 변경
domain -> global/entity로 이동
